Bare sshpass, rely on PATH to find it
Now with the new PATH inheritance in Acton's process module, it is
better to use a bare sshpass and rely on PATH to find it rather than
hard-coding a path.

Homebrew installs sshpass in /opt/homebrew, so practically this now
makes things work on MacOS and not just some linux distros.