Add SSH client
Extract the SSH client part from the NETCONF Client and place in its own
module such that it can be reused. Many things are also improved. We opt
for a simple design where on_connect is also handling connection errors
- it is very natural for on_connect to handle the initial connection
error but it's also quite natural that it handles such session errors
during the entire lifecycle.

We run sshpass / ssh in a more consistent fashion. We use -v verbose to
get more debug output that we use to e.g. determine that we are connected.