Revamp NETCONF client to use new SSH client
Get rid of the on_error callback, instead preferring to use on_connect.
The on_connect callback is already a natural place to handle failures
for connection setup. The idea of redirecting all errors to a kitchen
sink on_error callback might appear appealing since we can write happy
path in other callbacks, but it ultimately leads to hard to understand
error handling. So we extend on_connect with an error argument and
remove the on_error callback entirely. on_connect can thus be called at
any time, not just during initial connection setup but also later on, if
the connection fails.

There are new close and restart methods to simplify restarting the
client without having to replace it with a new instance.

More importantly, the internal SSH machinery have now been refactored to
use the new SSH client instead of directly running the ssh process. This
is a cleaner split and prepares us a bit for eventually replacing the
ssh_client with a libssh based one. The new ssh_client has working
on_connect callbacks etc, so everything works better like this.