Fix Buffer.find_bytes()
This is still a naive pattern matching algorithm in that it scans for
the entire pattern from the beginning at each index. But at least it is
addresses the problems of the previous implementation that incorrectly
matched the pattern with extra characters (as long as the pattern
sequence appeared) and did not handle overlapping patterns.