Add lock and unlock methods to NETCONF client
Add lock() and unlock() methods to Client class for datastore locking.
We parse the XML response and map to more idiomatic Acton style with an
error argument to the callback method, set to NetconfError if anything
goes wrong.

There's a _test_lock_unlock test case to test this, which has been
tested to work against sysrepo.