Refactor Buffer.find_bytes()
Use bytes.find(needle, start, end) for a more efficient search of the
first byte in the pattern. If found in a part, advance byte-by-byte and
search the rest of the pattern, possibly advancing to next part.
The last search position is cached and reused the next time, if the
pattern argument is the same.

Includes a performance test that replicates a real-world scenario to
avoid regressions in future.