Use serial RPC matching mode
This switches over the algorithm for matching rpc-replies with a
callback to the simple assumption of strict serialization in NETCONF.

NETCONF requires that RPC messages have a message-id set and NETCONF
servers are supposed to return the message-id, or really any XML
attribute set in the RPC message. We've relied on message-id for
matching up and finding the callback functions for RPC-replies. While
fairly simple and robust, it's turned out that some NETCONF servers
manage to mess up the message-id. We now switch over to assuming that
the callbacks are in order in the list of callbacks. Since NETCONF
guarantees strict sequence for messages, this should be fine. Hard to
say if there might be other quirks that breaks this - we will have to
see.