Add ietf-yang-library support to list_schemas
Add support for module discovery using ietf-yang-library (RFC 8525) in
addition to existing ietf-netconf-monitoring (RFC 6022). The
list_schemas action now automatically detects and uses yang-library when
advertised by the server.

The ietf-yang-library model defines a datastore schema to be a union of
module sets, and every module set is a collection of modules and
submodules, including the modules and submodules used for imports.

   +-----------+
   | datastore |
   +-----------+
        |
        | has a
        V
   +-----------+            +--------+                +------------+
   | datastore |  union of  | module |  consists of   | modules +  |
   |  schema   |----------->|  set   |--------------->| submodules |
   +-----------+            +--------+                +------------+

We attempt to retrieve the schema used by the "running" datastore, then
fetch the module set(s) and the respective list of modules. Extracting
this from the XML response is a bit convoluted since Acton xml.Node API
is very basic. Using adata & schema-driven parser would be nice here,
but are we ready to add the YANG library dependency to the NETCONF client?